+++ src/xtea/clip_read.py
@@ -355,7 +355,7 @@ class ClipReadInfo():
         for rcd in l_chrm_records:
             sf_clip_pos = self.working_folder + rcd[0] + xtea.global_values.CLIP_POS_SUFFIX
             sf_pub_pos=sf_pub_folder + rcd[0] + xtea.global_values.CLIP_POS_SUFFIX
-            if os.path.islink(sf_pub_pos)==True or os.path.isfile(sf_pub_pos)==True:
+            if (os.path.islink(sf_pub_pos)==True or os.path.isfile(sf_pub_pos)==True) and not xtea.global_values.KEEP_INT_FILES:
                 os.remove(sf_pub_pos)
             cmd="ln -s {0} {1}".format(sf_clip_pos, sf_pub_folder)
             Popen(cmd, shell=True, stdout=PIPE).communicate()
@@ -520,7 +520,8 @@ class ClipReadInfo():
                 with open(sf_clip_fq) as fin_clip:
                     for line in fin_clip:
                         fout_all.write(line)
-                os.remove(sf_clip_fq)#clean the temporary file
+                if not xtea.global_values.KEEP_INT_FILES:
+                    # os.remove(sf_clip_fq)#clean the temporary file
 ####
 
     def decrease_clip_freq(self, chrm, mpos, sflag, m_clip_freq):
